### 2024-09-05 #7184 @editorialbot
[PRE REVIEW]: cifkit: A user-friendly Python package for high-throughput CIF Analysis
**Submitting author:** @bobleesj (<a href="http://orcid.org/0000-0002-2367-3932">Sangjoon Lee</a>)
**Repository:** https://github.com/bobleesj/cifkit
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.0.2
