+++ paper.md
@@ -8,14 +8,13 @@ tags:
   - solid state chemistry
   - crystal structure
   - machine learning
-
 authors:
   - name: Sangjoon Lee
     orcid: 0000-0002-2367-3932
     corresponding: true
     affiliation: 1
   - name: Anton O. Oliynyk
-  - orcid: 0000-0003-0732-7340
+    orcid: 0000-0003-0732-7340
     affiliation: "2, 3"
 affiliations:
   - name: Department of Applied Physics and Applied Mathematics, Columbia University, New York, NY 10027, USA
@@ -24,7 +23,6 @@ affiliations:
     index: 2
   - name: Ph.D. Program in Chemistry, The Graduate Center of the City University of New York, New York, NY 10016, USA
     index: 3
-
 date: 29 August 2024
 bibliography: paper.bib
 ---
