### 2024-09-05 #7184 @bobleesj
Potential reviewers:

- [Matthew Evans](https://reviewers.joss.theoj.org/reviewers/1065)
- [Yuxing Fei](https://reviewers.joss.theoj.org/reviewers/4099)
