### 2024-09-05 #7184 @rkurchin
👋 @ml-evs, @espottesmith, and/or @lancekavalsky, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
