### 2024-09-09 #7205 @editorialbot
[REVIEW]: cifkit: A Python package for coordination geometry and atomic site analysis
**Submitting author:** @bobleesj (<a href="http://orcid.org/0000-0002-2367-3932">Sangjoon Lee</a>)
**Repository:** https://github.com/bobleesj/cifkit
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.6

@espottesmith,  @ml-evs, and @lancekavalsky, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
