### 2024-09-09 bobleesj/cifkit#27 @ml-evs
[JOSS review] Initial technical comments
Hi there, thanks for the submission to JOSS. Here are my initial technical comments; I'll post it here so you can control its visibility going forward, but I will link to it in the overall review issue with a few more general comments.

1. Unfortunately I am struggling to find many CIFs that actually work to test the functionality -- the error messages I am receiving are also not particularly illuminating. I tried:

    - [A random CIF from the COD](https://www.crystallography.net/cod/1100784.html) -- succesfully parsed but visualisation example failed for me.
    - [A random CIF from the Materials Project](https://next-gen.materialsproject.org/materials/mp-2282) -- failed to parse
    - A random CIF from the CSD ([CCDC 824282](http://webcsdoffsite.psds.ac.uk/Search?Ccdcid=824282))-- failed to parse
    - A CIF from the applications paper (Tyvanchuk et al) ([CCDC Number: 2294753](http://webcsdoffsite.psds.ac.uk/Search?Ccdcid=2294753&DatabaseToSearch=Published)) -- failed to parse

    These errors seem to relate to preprocessing functions outside of the CIF parser itself, but the errors will not be comprehensible for non-expert users. With the `CifEnsemble` class, all files get moved to `error_others` for me, and in fact, the COD CIF file that worked previously was somehow "corrupted" when being written out to `error_others`.

2. API documentation is lacking (function and class docstrings for the main public methods at a minimum). In particular, its neither clear which parts of CIF are supported (and if anything outside of the [core CIF dictionary](https://www.iucr.org/resources/cif/dictionaries) is supported), nor what to do in the face of errors, nor any discussion of symmetry.

3. I see tox is being used for pinned/testing environments. Given how quickly the Python ecosystem moves, I would suggest either lightly pinning the dependencies in `pyproject.toml` (to e.g., major versions) or at least mention in the documentation how one might install the tested pinned dependencies (i.e., potentially add it as an "extra" that you can install with `pip install .[strict]`, rather than needing another tool in tox). If this package is going to continue to be developed, you may also consider using a tool like [dependabot](https://docs.github.com/en/code-security/getting-started/dependabot-quickstart-guide) to automatically watch out for new breaking versions of your dependencies.
