### 2024-09-09 #7205 @ml-evs
This package exposes some convenience and utility functions on top of the [gemmi](https://joss.theoj.org/papers/10.21105/joss.04200) CIF parser, for handling coordination and bonding analysis, polyhedron visualisation and data cleaning for primarily inorganic crystal structures represented as CIFs. The headline is that I am not sure this package is sufficiently robust to the various intricacies of CIF (or at least, is not well-documented enough to cover them). I've raised some specific technical issues at https://github.com/bobleesj/cifkit/issues/27.

I'll try not to repeat @espottesmith's points (which I broadly agree with), so here's my first round of comments.

- While this package clearly has some useful tools (particularly the nice polyhedron visualisation), I fail to see how this tool is easier to use than pymatgen/ASE as a selling point. If it is targeted at people with limited programming/Python experience, then much more care must be taken in guiding them through the installation process (with virtual environments etc).
- Whilst "high throughput" is mentioned throughout the docs and paper, I see little in the way of e.g., assistance with parallelisation or batch processing (e.g., if I have 1m structures can I stream data without going via a file all the time?), nor any advanced error handling that could allow this package to be used in a highly automated way -- perhaps the authors could clarify what they mean with regards to high throughput in this case?
- I do not think it is appropriate to mention test coverage explicitly in the paper; although the package does appear to be well-tested with good CI, pinned dependencies and lots of test cases, this metric can be a red herring.

- There are a few glaring omissions in the references that could help provide useful background to readers:
    - The CIF framework itself is not appropriate cited (see https://www.iucr.org/resources/cif/cif2) -- the paper/documentation should also discuss which versions of the CIF standard are supported (it is a growing standard) with changes between v1 and v2.
    - Gemmi, the library used for the underlying CIF parsing, published in JOSS already https://joss.theoj.org/papers/10.21105/joss.04200
    - chemenv, used under the hood by pymatgen for coordination analysis https://journals.iucr.org/b/issues/2020/04/00/lo5066 (and subsequently much of the pre-existing literature for determining coordination environments).
    - Other dedicated CIF parsing projects supported by IUCr: [PyCIFRw](https://www.iucr.org/resources/cif/software/pycifrw) and [COD's CIF parser](http://wiki.crystallography.net/cod-tools/CIF-parser/) (with [pycodcif](https://pypi.org/project/pycodcif/2.4/) Python package) https://journals.iucr.org/j/issues/2016/01/00/po5052/index.html
    - There are many others listed at https://www.iucr.org/resources/cif/software
- I feel this submission is borderline on the "Substantial scholarly effort" front, given the guidelines at https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort I am not convinced by the text that the library was crucial in the three examples listed by the authors in the applications section; perhaps this could be expanded? Perhaps even the functionality from the other related packages by the author listed in the paper (SAF and CBA) could be exposed in the `cifcheck` namespace as (effectively) one package?
