### 2024-09-09 bobleesj/cifkit#27 @bobleesj
Thank you, @ml-evs, for your time spent testing and posting the issues here.

1. I will double-check the file sources from the links above. I have recently supported and tested against .cif files sourced from COD, ICSD, PCD, and Materials Studio (Support various .cif formats for v 1.0.3 (COD, ICSD, PCD, Materials Studio) #24), but I will verify them again.

2. I will provide API documentation in the next iteration.

3. I agree with your point about having recently removed the strict pinning as reflected in the current version: [pyproject.toml](https://github.com/bobleesj/cifkit/blob/main/pyproject.toml). I will also explore dependabot and share my feedback.
