### 2024-09-15 bobleesj/cifkit#31 @lancekavalsky
Ingestion works with Cif but not CifEnsemble
**Checklist**
- [x] Are you using the latest stable version?

**Describe the bug**
I have noticed that while `Cif` can ingest the Materials Project `cif`s I tried (e.g. [mp-19294](https://next-gen.materialsproject.org/materials/mp-19294?chemsys=Li-P-O-Fe), [mp-804](https://next-gen.materialsproject.org/materials/mp-804?chemsys=Ga-N)), `CifEnsemble` logs these as `error_others` when placed in a directory.

**To Reproduce**
```
>>> from cifkit import CifEnsemble
>>> # where mp_test is a folder containing the MP cif files
>>> ensemble = CifEnsemble("mp_test")
Error processing mp_test/LiFeP2O7.cif: mp_test/LiFeP2O7.cif:1:0(0): expected block header (data_)
Error processing mp_test/GaN.cif: mp_test/GaN.cif:1:0(0): expected block header (data_)

CIF Preprocessing in mp_test begun...

Preprocessing mp_test/LiFeP2O7.cif (1/2)
File LiFeP2O7.cif moved to 'error_others' due to: mp_test/LiFeP2O7.cif:1:0(0): expected block header (data_)
Preprocessing mp_test/GaN.cif (2/2)
File GaN.cif moved to 'error_others' due to: mp_test/GaN.cif:1:0(0): expected block header (data_)

SUMMARY
# of files moved to 'error_operations' folder: 0
# of files moved to 'error_duplicate_labels' folder: 0
# of files moved to 'error_wrong_loop_value' folder: 0
# of files moved to 'error_coords' folder: 0
# of files moved to 'error_invalid_label' folder: 0
# of files moved to 'error_others' folder: 2

Initializing 0 Cif objects...
Finished initialization!
```

**Expected behavior**
Ingestion of the CIF files into a `CifEnsemble` object
