### 2024-09-15 #7205 @lancekavalsky
In this work, the `cifkit` package is presented to streamline the manipulation of `cif` structure files en masse through various utility functions including coordination analysis, visualization, and filtering by property. Overall, the package is cleanly written with a largely intuitive API from a user standpoint, with some useful features for easy visualization and error-handling upon ingestion.

Code comments:

- For `io` related methods, e.g. the ones that generate histograms, more clarity is needed regarding where they write things by default. Running the histogram tutorial from the documentation, it wrote the histograms to a folder deep in my conda environment site-packages, which would likely not be intuitive to many users (particularly as this package is presented as being catered to users with less coding experience) and may cause issues on shared resources.

- I will echo the previous comment that more docstrings would be invaluable in helping code clarity. In particular, I would urge adopting a standardized approach to this, such as [Numpy](https://numpydoc.readthedocs.io/en/latest/format.html), to be more in line with community standards

- I have opened a technical issue I encountered [here](https://github.com/bobleesj/cifkit/issues/31)

Documentation comments:

- As the core classes for this package are `Cif` and `CifEnsemble`, more explicit explanations as to the inputs and parameters would be helpful -- especially for `CifEnsemble`. For example, unless I'm missing it, a comprehensive list of everything that the`preprocess` parameter triggers is not mentioned in the documentation.

- In the documentation there are a couple instances of general clean-up required. One example is the first box on `Getting Started` uses a CIF method `ensemble.cif_folder_path` which gives an error when run. Another example is under the `CIF` specific documentation which refers to a `README.md` for complete documentation, but it is unclear where this file is located (since that info doesn't appear to be the main README in the repo?).

> There are options in the Cif class to use either the by_d_min_method or by_best_methods. Please refer to the README.md for complete documentation.

- This is relatively minor, but on the documentation website it wasn't immediately clear to me what would be contained under the `Notebooks` tab at the top. Given this outlines several of the core functionalities of `cifkit`, I would consider renaming it to something more descriptive.

- The contributing guidelines at this stage are somewhat vague. For example, is a minimum code coverage required for added features?

Paper comments:

- In the summary it mentions that this package is designed to process datasets on "the order of tens of thousands". It is not clear to me where this is coming from and what exactly causes the bottleneck for going beyond this estimate. Details regarding what determined this limitation would be helpful to judge high-throughput performance

- The examples in the paper are presented with limited explanation as to what they are showing. While the comments help in the second example, and some of the methods are self-explanatory by their naming, more comments would help cement the clarity here.

Overall, this work would make a great addition to JOSS pending the minor revisions described above.
