### 2024-09-19 c25be06
Delete news/ci-test.rst
