+++ news/ci-test.rst
@@ -1,23 +1,0 @@
-**Added:**
-
-* <news item>
-
-**Changed:**
-
-* Testing puporses
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* <news item>
-
-**Fixed:**
-
-* <news item>
-
-**Security:**
-
-* <news item>
