+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -16,9 +16,7 @@ A clear and concise description of what the bug is.
 
 **To Reproduce**
 Steps to reproduce the behavior:
-1. 
-2. 
-3. 
+1. Go to ...
 
 **Expected behavior**
 A clear and concise description of what you expected to happen.
