### 2024-09-19 f13da18
Update bug_report.md
