+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -16,7 +16,7 @@ A clear and concise description of what the bug is.
 
 **To Reproduce**
 Steps to reproduce the behavior:
-1. 
+1. Go to
 2. 
 3. 
 
