### 2024-09-21 7ec8175
Add config in pyproject.toml
