### 2024-09-21 6ac9b0b
Add version to init
