+++ src/cifkit/version.py
@@ -1,3 +1,0 @@
-from importlib.metadata import version
-
-__version__ = version("cifkit")
