### 2024-09-21 ce35f9a
Ensure __version__ exists
