+++ tests/test__init__.py
@@ -1,0 +1,9 @@
+import pytest
+
+import cifkit
+
+
+@pytest.mark.now
+def test_version():
+    assert cifkit.__version__
+    assert cifkit.__version__ != "0.0.0"
