+++ .github/workflows/wheels.yml
@@ -63,4 +63,4 @@ jobs:
 
       - uses: pypa/gh-action-pypi-publish@release/v1
         with:
-          # To test: repository-url: https://test.pypi.org/legacy/
+          repository-url: https://test.pypi.org/legacy/
