### 2024-10-07 8edbef2
Add running script for rc only
