+++ .github/workflows/build-wheels-publish.yml
@@ -28,8 +28,16 @@ jobs:
           name: Packages
           path: dist
 
-      - name: Publish package distributions to TestPyPI
+      # Upload to PyPI
+      - name: Publish package distributions to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
           repository-url: https://test.pypi.org/legacy/
           verbose: true
+      
+      # This step runs for "rc" tags
+      - name: Run specific script for RC tags
+        if: contains(github.ref, 'rc')
+        run: |
+          echo "Running RC-specific script..."
+          ./scripts/rc-specific-script.sh  # Replace with your actual script
