+++ .github/workflows/build-wheels-publish.yml
@@ -11,7 +11,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
         with:
-          fetch-depth: 0
+          ref: ${{ github.head_ref }}
       - uses: hynek/build-and-inspect-python-package@v2
 
   pypi-publish:
@@ -38,19 +38,23 @@ jobs:
   update-changelog:
     needs: [dist, pypi-publish]
     runs-on: ubuntu-latest
+    permissions:
+      # Give the default GITHUB_TOKEN write permission to commit and push the
+      # added or changed files to the repository.
+      contents: write
     steps:
       - uses: actions/checkout@v4
         with:
-          fetch-depth: 0
-      - name: Run specific script for non-RC tags
-        if: "!contains(github.ref, 'rc')"
-        run: |
-          python .github/workflows/news-changelog.py
-      - name: Commit and push if there are changes
-        run: |
-          git add -A
-          git commit -m "Update changelog"
-          git push origin HEAD:${{ github.ref }}
+          ref: ${{ github.head_ref }}
+      - uses: stefanzweifel/git-auto-commit-action@v5
+        with:
+          commit_message: update changelog
+      # - name: Run specific script for non-RC tags
+      #   if: "!contains(github.ref, 'rc')"
+      #   run: |
+      #     python .github/workflows/news-changelog.py
+      # - name: Commit and push if there are changes
+      #   uses: stefanzweifel/git-auto-commit-action@v5
 
           
       # - name: Run specific script for RC tags
