### 2024-10-07 3f7bd9d
Apply changelog script in workflow
