+++ news/delete-mypi.rst
@@ -1,23 +1,0 @@
-**Added:**
-
-* <news item>
-
-**Changed:**
-
-* <news item>
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* mypy.ini file
-
-**Fixed:**
-
-* <news item>
-
-**Security:**
-
-* <news item>
