+++ news/news-ci.rst
@@ -1,23 +1,0 @@
-**Added:**
-
-* <news item>
-
-**Changed:**
-
-* <news item>
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* <news item>
-
-**Fixed:**
-
-* News CI bug with existing news *rst files
-
-**Security:**
-
-* <news item>
