+++ news/test.rst
@@ -1,0 +1,23 @@
+**Added:**
+
+* addeed 1
+**Changed:**
+
+* changed 1
+* changed 2
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* <news item>
+
+**Security:**
+
+* <news item>
