+++ CHANGELOG.rst
@@ -7,6 +7,10 @@ Release Notes
 3.2.2
 =====
 
+
+3.2.2
+=====
+
 **Added:**
 
 * addeed 1
