### 2024-10-07 967c072
Add proper condition for rc
