+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ Release Notes
 3.2.2
 =====
 
+
+3.2.2
+=====
+
 **Added:**
 
 * addeed 1
