### 2024-10-07 1b956ff
Add automatic release note for rc
