+++ .github/workflows/build-wheels-publish.yml
@@ -4,31 +4,31 @@ on:
   push:
     tags:
       - '*' 
+permissions:
+  contents: write
 jobs:
   build-package:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
         with:
-          ref: ${{ github.head_ref }}
+          ref: ${{ github.ref }}
       - uses: hynek/build-and-inspect-python-package@v2
 
   update-changelog:
     needs: [build-package]
     runs-on: ubuntu-latest
-    permissions:
-      contents: write
     steps:
       - name: Checkout the repository
         uses: actions/checkout@v4
         with:
-          ref: ${{ github.head_ref }}
+          ref: ${{ github.ref }}
 
-      - name: Run the changelog script for non-rc tags
+      - name: Run the changelog script for full release (no rc)
         if: "!contains(github.ref, 'rc')"
         run: python .github/workflows/news-changelog.py
 
-      - name: Automatically commit changes for non-rc tags
+      - name: Automatically commit changes full release (no rc)
         if: "!contains(github.ref, 'rc')"
         uses: stefanzweifel/git-auto-commit-action@v5
         with:
@@ -36,7 +36,7 @@ jobs:
           branch: pypi-build  # FIXME: change to main
    
       # - name: Run specific script for RC tags
-      #   if: contains(github.ref, 'rc')
+      #   if: contains(githgitub.ref, 'rc')
 
   pypi-publish:
     needs: [build-package, update-changelog]
@@ -52,10 +52,21 @@ jobs:
           name: Packages
           path: dist
 
-      # Upload to PyPI
       - name: Publish package distributions to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
           repository-url: https://test.pypi.org/legacy/
           verbose: true
-      
+
+  create-release:
+    needs: [update-changelog]
+    runs-on: ubuntu-latest
+    steps:
+      - name: Create Release for Full Versions
+        if: contains(github.ref, 'rc')
+        uses: softprops/action-gh-release@v2
+        with:
+          files: 'dist/*'
+          body: "Full release details and changelog."
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
