+++ .github/workflows/build-wheels-publish.yml
@@ -65,5 +65,12 @@ jobs:
       - name: Create Release for Full Versions
         if: contains(github.ref, 'rc')
         uses: softprops/action-gh-release@v2
+        with:
+          tag_name: ${{ github.ref }}
+          title: ${{ github.ref }}
+          body: ${{ steps.update-changelog.outputs.changelog }}
+          token: ${{ secrets.GITHUB_TOKEN }}
+          draft: true
+          prerelease: true
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
