### 2024-10-07 b0f5be7
move default setting in gh release
