### 2024-10-07 edbf149
generate release notes
