+++ .github/workflows/build-wheels-publish.yml
@@ -68,5 +68,6 @@ jobs:
         with:
           draft: true
           prerelease: true
+          generate_release_notes: true
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
