+++ news/TEMPLATE copy 2.rst
@@ -1,0 +1,24 @@
+**Added:**
+
+* <news item>
+
+**Changed:**
+
+* <news item>
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* mypy.ini file
+
+
+**Fixed:**
+
+* News CI bug with existing news *rst files
+
+**Security:**
+
+* <news item>
