+++ .github/workflows/build-wheels-publish.yml
@@ -17,6 +17,7 @@ jobs:
 
   update-changelog:
     needs: [build-package]
+    if: "!contains(github.ref, 'rc')"
     runs-on: ubuntu-latest
     steps:
       - name: Checkout the repository
@@ -25,19 +26,14 @@ jobs:
           ref: ${{ github.ref }}
 
       - name: Run the changelog script for full release (no rc)
-        if: "!contains(github.ref, 'rc')"
-        run: python .github/workflows/news-changelog.py
+        run: python .github/workflows/news-changelog.py "${{ github.ref }}"
 
       - name: Automatically commit changes full release (no rc)
-        if: "!contains(github.ref, 'rc')"
         uses: stefanzweifel/git-auto-commit-action@v5
         with:
           commit_message: update changelog
           branch: pypi-build  # FIXME: change to main
-   
-      # - name: Run specific script for RC tags
-      #   if: contains(githgitub.ref, 'rc')
-
+  
   pypi-publish:
     needs: [build-package, update-changelog]
     environment:
@@ -58,16 +54,23 @@ jobs:
           repository-url: https://test.pypi.org/legacy/
           verbose: true
 
-  create-release:
-    needs: [update-changelog]
+  github-pre-release:
+    if: contains(github.ref, 'rc')
     runs-on: ubuntu-latest
     steps:
-      - name: Create Release for Full Versions
-        if: contains(github.ref, 'rc')
+      - name: Create pre-release for rc versions
         uses: softprops/action-gh-release@v2
         with:
           draft: true
           prerelease: true
           generate_release_notes: true
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
+  github-release:
+    if: "!contains(github.ref, 'rc')"
+    runs-on: ubuntu-latest
+    steps:
+      - name: Create full release 
+        uses: softprops/action-gh-release@v2
+        with:
+          draft: true
+          
