+++ CHANGELOG.rst
@@ -4,6 +4,10 @@ Release Notes
 
 .. current developments
 
+1.0.8
+=====
+
+
 1.0.7
 =====
 
