### 2024-10-07 c3aa14f
capture latest news as gh env variable
