+++ changelog.txt
@@ -1,1 +1,15 @@
-3.2.2
+1.0.8
+=====
+
+**Added:**
+
+* Test 1
+
+**Fixed:**
+
+* News CI bug with existing news *rst files
+
+**Removed:**
+
+* mypy.ini file
+
