+++ news/TEMPLATE copy 2.rst
@@ -1,23 +1,0 @@
-**Added:**
-
-* Test
-
-**Changed:**
-
-* <news item>
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* <news item>
-
-**Fixed:**
-
-* <news item>
-
-**Security:**
-
-* <news item>
