+++ news/TEMPLATE copy.rst
@@ -1,23 +1,0 @@
-**Added:**
-
-* <news item>
-
-**Changed:**
-
-* Test3
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* <news item>
-
-**Fixed:**
-
-* <news item>
-
-**Security:**
-
-* <news item>
