+++ .github/workflows/build-wheels-publish.yml
@@ -76,7 +76,4 @@ jobs:
         with:
           body_path: CHANGELOG.txt
           draft: true
-          # note you'll typically need to create a personal access token
-          # with permissions to create releases in the other repo
-          # token: ${{ secrets.CUSTOM_GITHUB_TOKEN }}
           
