+++ .github/workflows/build-wheels-publish.yml
@@ -73,8 +73,13 @@ jobs:
         uses: actions/checkout@v4
         with:
           ref: ${{ github.ref }}
-      - name: Generate release notes for full release (no rc)
+      - name: Generate release notes for full release for full release
         run: python .github/workflows/get-latest-changelog.py "${{ github.ref_name }}"
+      - name: Commit the latest release note for full release
+        uses: stefanzweifel/git-auto-commit-action@v5
+        with:
+          commit_message: update changelog
+          branch: pypi-build  # FIXME: change to main
       - name: Release
         uses: softprops/action-gh-release@v2
         with:
