+++ news/TEMPLATE copy.rst
@@ -1,23 +1,0 @@
-**Added:**
-
-* <news item>
-
-**Changed:**
-
-* <news item>
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* 4
-
-**Fixed:**
-
-* <news item>
-
-**Security:**
-
-* <news item>
