### 2024-10-07 c408171
Add print for debug latest changelog
