+++ news/TEMPLATE copy.rst
@@ -1,0 +1,23 @@
+**Added:**
+
+* 2
+
+**Changed:**
+
+* <news item>
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* <news item>
+
+**Security:**
+
+* <news item>
