+++ CHANGELOG.txt
@@ -1,0 +1,15 @@
+1.0.13
+=====
+
+**Added:**
+
+* 1
+
+**Changed:**
+
+* 2
+
+**Deprecated:**
+
+* 3
+
