### 2024-10-07 d353cdd
require changelog update before get latest changelog
