+++ .github/workflows/build-wheels-publish.yml
@@ -66,6 +66,7 @@ jobs:
           generate_release_notes: true
 
   github-release:
+    needs: [update-changelog]
     if: "!contains(github.ref, 'rc')"
     runs-on: ubuntu-latest
     steps:
