+++ .github/workflows/build-wheels-publish.yml
@@ -77,7 +77,11 @@ jobs:
       - name: Generate release notes for full release for full release
         run: python .github/workflows/get-latest-changelog.py "${{ github.ref_name }}"
       - name: Check if CHANGELOG.txt was generated
-        run: ls -la  # This will show you if CHANGELOG.txt was generated and where it is located
+        run: |
+          ls -la
+          cat CHANGELOG.rst
+          cat CHANGELOG.txt
+
       - name: Release
         uses: softprops/action-gh-release@v2
         with:
