+++ news/TEMPLATE copy 2.rst
@@ -1,0 +1,23 @@
+**Added:**
+
+* 1
+
+**Changed:**
+
+* 2
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* <news item>
+
+**Security:**
+
+* <news item>
