### 2024-10-07 432b294
Refactor python code for changelog
