+++ CHANGELOG.txt
@@ -1,8 +1,13 @@
-1.0.16
-=====
 
 **Added:**
 
-* 2
 * 1
+
+**Changed:**
+
+* 2
+
+**Deprecated:**
+
+* 3
 
