### 2024-10-07 #7205 @bobleesj
@rkurchin Thanks for checking in with me. I have made progress and addressed about halfway through. Would you please allow me to respond by the third week of October?
