### 2024-10-08 f37e307
Add template for release
