### 2024-10-08 d182e4b
Add permission id-token at template level
