### 2024-10-09 56f77f6
remove news check
