+++ .github/workflows/build-wheel-release-publish.yml
@@ -1,10 +1,11 @@
 name: Release
 
 on:
-  workflow_dispatch:
-  push:
-    tags:
-      - '*'
+  pull_request:
+  # workflow_dispatch:
+  # push:
+  #   tags:
+  #     - '*'
 
 permissions:
   id-token: write
