+++ .github/workflows/build-wheel-release-publish.yml
@@ -1,7 +1,6 @@
 name: Release
 
 on:
-  pull_request:
   workflow_dispatch:
   push:
     tags:
