### 2024-10-26 f908180
Update issue and PR templates
