+++ .github/pull_request_template.md
@@ -1,21 +1,0 @@
-## What type of PR is this? (check all applicable)
-
-- [ ] Refactor
-- [ ] Feature
-- [ ] Bug Fix
-- [ ] Optimization
-- [ ] Documentation Update
-
-## Description (Screenshots, files, etc)
-
-## Checklist
-
-- [ ] Are the tests passing?
-- [ ] If it's a new feature, have tests been written?
-
-## Added to documentation?
-
-- [ ] README.md
-- [ ] Official documentation
-- [ ] Google Codelab
-- [ ] No documentation needed
