+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,27 +1,0 @@
----
-name: Bug report
-about: Create a report to help us improve
-title: ""
-labels: ""
-assignees: ""
----
-
-**Checklist**
-
-- [ ] Are you using the latest stable version?
-
-**Describe the bug** A clear and concise description of what the bug is.
-
-**To Reproduce** Steps to reproduce the behavior:
-
-1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
-
-**Expected behavior** A clear and concise description of what you expected to
-happen.
-
-**Screenshots** If applicable, add screenshots to help explain your problem.
-
-**How to fix (optional)**
