+++ .github/workflows/build-wheel-release-upload.yml
@@ -15,3 +15,4 @@ jobs:
       project: cifkit
     secrets:
       PYPI_TOKEN: ${{ secrets.PYPI_TOKEN }}
+      PAT_TOKEN: ${{ secrets.PAT_TOKEN }}
