### 2024-10-26 03f9ddc
Skip pre-commit for cif files
