+++ news/pre-commit.rst
@@ -1,0 +1,24 @@
+**Added:**
+
+* pre-commit hook for each pull request
+* Support for Python 3.13
+
+**Changed:**
+
+* <news item>
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* Support for Python 3.10
+
+**Fixed:**
+
+* Simpler issue templates used in Billinge Group
+
+**Security:**
+
+* <news item>
