### 2024-10-27 b4c0e27
Add env.yml conda file
