+++ environment.yml
@@ -1,0 +1,6 @@
+name: diffpy.snmf
+channels:
+  - conda-forge
+dependencies:
+  - python=3
+  - pip
