+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "cifkit"
 dynamic=['version', 'dependencies']
-requires-python = ">=3.11, <3.14"
+requires-python = ">=3.10, <3.13"
 classifiers = [
         'Development Status :: 5 - Production/Stable',
         'Environment :: Console',
