### 2024-10-27 c2a1a66
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
