### 2024-10-27 411e698
remove centralized workflow for CI run
