+++ .github/workflows/tests-on-pr.yml
@@ -1,16 +1,36 @@
-name: Tests on PR
+name: CI
 
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-  workflow_dispatch:
+on: [push, pull_request]
 
 jobs:
-  validate:
-    uses: bobleesj/release-scripts/.github/workflows/_tests-on-pr.yml@v0
-    with:
-      project: cifkit
-      c_extension: false
-      headless: false
+  build-linux:
+    runs-on: ubuntu-latest
+    strategy:
+      max-parallel: 5
+      matrix:
+        python-version: ["3.10", "3.11", "3.12"]
+
+    steps:
+    - uses: actions/checkout@v4
+
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+
+    - name: Create virtual environment and install dependencies
+      run: |
+        python -m venv venv
+        source venv/bin/activate
+        pip install .
+
+    - name: Test with pytest and generate coverage report
+      run: |
+        source venv/bin/activate
+        pip install pytest pytest-cov
+        python -m pytest -m "not pyvista" --cov=./ --cov-report=xml
+
+    - name: Upload coverage reports to Codecov
+      uses: codecov/codecov-action@v4.0.1
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
