+++ .github/workflows/matrix-and-codecov-on-merge-to-main.yml
@@ -1,21 +1,0 @@
-name: CI
-
-on:
-  push:
-    branches:
-      - main
-  release:
-    types:
-      - prereleased
-      - published
-  workflow_dispatch:
-
-jobs:
-  coverage:
-    uses: Billingegroup/release-scripts/.github/workflows/_matrix-and-codecov-on-merge-to-main.yml@v0
-    with:
-      project: cifkit
-      c_extension: false
-      headless: false
-    secrets:
-      CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
