+++ environment.yml
@@ -1,6 +1,0 @@
-name: diffpy.snmf
-channels:
-  - conda-forge
-dependencies:
-  - python=3
-  - pip
