### 2024-10-27 ba14829
Run CI on push to main and pr
