### 2024-10-27 adff820
run edit cif file based on db in a cif_editor.py
