### 2024-10-27 291613f
add MS support in cif_editor and sourcer
