+++ src/cifkit/utils/cif_sourcer.py
@@ -7,6 +7,7 @@ def get_cif_db_source(file_path):
         "ICSD": "_database_code_ICSD",
         "MS": "'Materials Studio'",
         "PCD": "#_database_code_PCD",
+        "MP": "# generated using pymatgen",
     }
 
     if os.path.exists(file_path) and file_path.endswith(".cif"):
