### 2024-10-27 0dd5e4c
add init tests for MS and other db sourced cif files
