+++ tests/core/coordination/test_bond_distance.py
@@ -1,4 +1,6 @@
-from cifkit.coordination.bond_distance import get_shortest_distance_per_bond_pair
+from cifkit.coordination.bond_distance import (
+    get_shortest_distance_per_bond_pair,
+)
 
 
 def test_get_minimum_dist_per_bond_pair(
