+++ news/MP.rst
@@ -1,0 +1,23 @@
+**Added:**
+
+* CifEnsemble support for ICSD, COD, MP files
+
+**Changed:**
+
+* <news item>
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* Preprocess .cif files in CifEnsemble before initializing into CIF objects
+
+**Security:**
+
+* <news item>
