### 2024-10-27 57dd584
Add CCSD test to cif ensemble
