### 2024-10-27 bf97c7e
Apply docstring to src and format doc strings
