+++ src/cifkit/data/radius_handler.py
@@ -2,13 +2,12 @@ import numpy as np
 
 from cifkit.data.radius import get_radius_data
 from cifkit.data.radius_optimization import get_refined_CIF_radius
+from cifkit.utils.unit import round_dict_values
 
 
 def get_is_radius_data_available(elements: list[str]) -> bool:
-    """
-    Check if both CIF and Pauling radius data are available
-    for each element in the list.
-    """
+    """Check if both CIF and Pauling radius data are available for each element in
+    the list."""
     data = get_radius_data()
     for element in elements:
         element_data = data.get(element, {})
@@ -21,9 +20,7 @@ def get_is_radius_data_available(elements: list[str]) -> bool:
 
 
 def get_CIF_pauling_radius(elements: list[str]) -> dict:
-    """
-    Return CIF and Pualing data for a list of elements
-    """
+    """Return CIF and Pualing data for a list of elements."""
     data = get_radius_data()
     radii = {}
     for atom in elements:
@@ -35,10 +32,8 @@ def get_CIF_pauling_radius(elements: list[str]) -> dict:
     return radii
 
 
-def get_radius_values_per_element(elements, shortest_bond_distances) -> dict:
-    """
-    Merge CIF and Pauling radius data with CIF refined radius data.
-    """
+def get_radius_values_per_element(elements: list[str], shortest_bond_distances) -> dict:
+    """Merge CIF and Pauling radius data with CIF refined radius data."""
     is_radius_data_available = get_is_radius_data_available(elements)
 
     if not is_radius_data_available:
@@ -59,13 +54,11 @@ def get_radius_values_per_element(elements, shortest_bond_distances) -> dict:
             ],
         }
 
-    return combined_radii
+    return round_dict_values(combined_radii)
 
 
-def compute_radius_sum(radius_values, is_radius_data_available: bool):
-    """
-    Compute the sum of two radii.
-    """
+def compute_radius_sum(radius_values: dict[str: dict[str:float]], is_radius_data_available: bool):
+    """Compute the sum of two radii."""
 
     if not is_radius_data_available:
         return None
