+++ src/cifkit/models/cif.py
@@ -59,8 +59,9 @@ from cifkit.utils.cif_parser import (
 
 # Identify .cif database source
 from cifkit.utils.cif_sourcer import get_cif_db_source
+
+# Utility
 from cifkit.utils.log_messages import CifLog
-from cifkit.utils.unit import round_dict_values
 
 
 def ensure_connections(func):
@@ -85,29 +86,92 @@ class Cif:
     def __init__(
         self, file_path: str, is_formatted=False, logging_enabled=False
     ) -> None:
-        """_summary_
+        """Initialize an object from a .cif file.
 
         Parameters
         ----------
         file_path : str
-            _description_
+            Path to the .cif file.
         is_formatted : bool, optional
-            _description_, by default False
+            If False, preprocesses the .cif file for compatibility with the gemmi library, by default False.
         logging_enabled : bool, optional
-            _description_, by default False
+            Enables logging for initialization and distance calculations, by default False.
+
+
+        Attributes
+        ----------
+        file_path : str
+            Path to the CIF file from which data is loaded.
+        logging_enabled : bool
+            If True, enables detailed logging for initialization and distancecalculations.
+        file_name : str
+            Base name of the CIF file, extracted from `file_path`.
+        file_name_without_ext : str
+            File name without its extension, useful for referencing or generating derivative files.
+        db_source : str
+            Source database (ICSD, MP, CCDC, PCD, etc.) from which the CIF file originates, determined at runtime.
+        unitcell_lengths : list[float]
+            List of unit cell lengths for the crystal structure, typically in Angstroms.
+        unitcell_angles : list[float]
+            List of unit cell angles in radians ordred by alpha, beta, gamma.
+        site_labels : list[str]
+            List of all unique atomic site labels.
+        unique_elements : set[str]
+            Set of unique chemical elements present in the cif file.
+        atom_site_info : dict[str, any]
+            Dictionary containing detailed element, site occupancy, fractional coords, symmetry, and multiplicity
+            information for each atomic site
+        composition_type : int
+            Number of unique elements present in the .cif file. 1 for unary, 2 for binary, etc.
+        tag : str
+            Additional tag associated with the CIF data, parsed from the third line of PCD .cif files.
+        bond_pairs : set[tuple[str, str]]
+            Set of tuples representing bonded pairs of elements.
+        site_label_pairs : set[tuple[str, str]]
+            Set of tuples representing pairs of atomic site labels
+        bond_pairs_sorted_by_mendeleev : set[tuple[str, str]]
+            Set of bonded pairs sorted according to Mendeleev Numbers.
+        site_label_pairs_sorted_by_mendeleev : set[tuple[str, str]]
+            Set of site label pairs sorted by Mendeleev Numbers.
+       site_mixing_type : str
+            Descriptor of the mixing type, categorized into deficiency_atomic_mixing, full_occupancy_atomic_mixing,
+            deficiency_without_atomic_mixing, and full_occupancy.
+        is_radius_data_available : bool
+            Indicates whether data about Pauling and CIF atomic radii are available for the elements in the .cif.
+        mixing_info_per_label_pair : dict
+            Dictionary mapping pairs of labels to their mixing information.
+        mixing_info_per_label_pair_sorted_by_mendeleev : dict
+            Same as `mixing_info_per_label_pair`, but sorted according to Mendeleev's order.
+        unitcell_points : list[list[tuple[float, float, float, str]]]
+            List of points defining the unit cell; each point contains fractional coordinates and a site label.
+        supercell_points : list[list[tuple[float, float, float, str]]]
+            List of points defining the supercell of the cell, with translations of ±1, ±1, ±1 from the unit cell.
+        unitcell_atom_count : int
+            Total count of atoms within the unit cell.
[…]
