+++ src/cifkit/preprocessors/supercell_util.py
@@ -1,5 +1,3 @@
 def get_cell_atom_count(supercell_points) -> int:
-    """
-    Count the number of atoms in the cell.
-    """
+    """Count the number of atoms in the cell."""
     return len(supercell_points)

[…]
