### 2024-10-27 2852cd1
Apply docstring fix in tests and fix typos
