### 2024-10-27 5d7a805
Apply static method to get file path in example.py
