### 2024-10-27 9c21b2f
Revert "Apply static method to get file path in example.py"
This reverts commit 5d7a805ccc2b90857ac6c0e1c46ada59294e9f76.
