+++ news/codespell.rst
@@ -1,24 +1,0 @@
-**Added:**
-
-* Added codespell to the pre-commit hooks to catch common spelling mistakes
-* Add docstrings to cif.py
-
-**Changed:**
-
-* <news item>
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* <news item>
-
-**Fixed:**
-
-* <news item>
-
-**Security:**
-
-* <news item>
