+++ src/cifkit/models/cif.py
@@ -86,77 +86,89 @@ class Cif:
     def __init__(
         self, file_path: str, is_formatted=False, logging_enabled=False
     ) -> None:
-        """Initialize an object from a .cif file.
+        """Initializes an object from a .cif file.
 
         Parameters
         ----------
         file_path : str
             Path to the .cif file.
         is_formatted : bool, optional
-            If False, preprocesses the .cif file for compatibility with the gemmi library, by default False.
+            If False, preprocess the .cif file to ensure compatibility with the
+            gemmi library. Default is False.
         logging_enabled : bool, optional
-            Enables logging for initialization and distance calculations, by default False.
-
+            Enables detailed logging during initialization and for distance
+            calculations. Default is False.
 
         Attributes
         ----------
         file_path : str
             Path to the CIF file from which data is loaded.
         logging_enabled : bool
-            If True, enables detailed logging for initialization and distancecalculations.
+            Enables detailed logging for initialization and distance
+            alculations if set to True.
         file_name : str
             Base name of the CIF file, extracted from `file_path`.
         file_name_without_ext : str
-            File name without its extension, useful for referencing or generating derivative files.
+            File name without its extension, useful for referencing or
+            generating derivative files.
         db_source : str
-            Source database (ICSD, MP, CCDC, PCD, etc.) from which the CIF file originates, determined at runtime.
+            Source database (e.g., ICSD, MP, CCDC, PCD) from which the CIF file
+            originates, determined at runtime.
         unitcell_lengths : list[float]
-            List of unit cell lengths for the crystal structure, typically in Angstroms.
+            List of unit cell lengths for the crystal structure, typically in
+            Angstroms.
         unitcell_angles : list[float]
-            List of unit cell angles in radians ordred by alpha, beta, gamma.
+            List of unit cell angles in radians, ordered by alpha, beta, gamma.
         site_labels : list[str]
-            List of all unique atomic site labels.
+            Lists all unique atomic site labels.
         unique_elements : set[str]
-            Set of unique chemical elements present in the cif file.
+            Set of unique chemical elements present in the CIF file.
         atom_site_info : dict[str, any]
-            Dictionary containing detailed element, site occupancy, fractional coords, symmetry, and multiplicity
-            information for each atomic site
+            Dictionary containing detailed information about each atomic site
+            including element, site occupancy,
+            fractional coordinates, symmetry, and multiplicity.
         composition_type : int
-            Number of unique elements present in the .cif file. 1 for unary, 2 for binary, etc.
+            Number of unique elements present in the .cif file, e.g., 1 for
+            unary, 2 for binary, etc.
         tag : str
-            Additional tag associated with the CIF data, parsed from the third line of PCD .cif files.
+            Additional tag associated with the CIF data, parsed from the third
+            line of PCD .cif files.
         bond_pairs : set[tuple[str, str]]
             Set of tuples representing bonded pairs of elements.
         site_label_pairs : set[tuple[str, str]]
-            Set of tuples representing pairs of atomic site labels
+            Set of tuples representing pairs of atomic site labels.
         bond_pairs_sorted_by_mendeleev : set[tuple[str, str]]
             Set of bonded pairs sorted according to Mendeleev Numbers.
         site_label_pairs_sorted_by_mendeleev : set[tuple[str, str]]
             Set of site label pairs sorted by Mendeleev Numbers.
-       site_mixing_type : str
-            Descriptor of the mixing type, categorized into deficiency_atomic_mixing, full_occupancy_atomic_mixing,
-            deficiency_without_atomic_mixing, and full_occupancy.
[…]
