+++ src/cifkit/coordination/connection.py
@@ -1,7 +1,7 @@
 def get_CN_connections_by_best_methods(
     best_methods, connections: dict
 ) -> dict:
-    """Retrieve connections limited by the number of vertices (CN_value) for each
+    """Retrieve connections limited by the number of vertices (CN) for each
     label."""
     CN_connections = {}
 
