### 2024-10-27 bobleesj/cifkit#53 @bobleesj
Refactor generating histograms based on attributes in CifEnsemble
### Problem

```
 # FIXME: refactor this section to maintain DRY principle

    def generate_structure_histogram(self, display=False, output_dir=None):
        plot_histogram(
            "structure",
            self.structure_stats,
            self.dir_path,
            display,
            output_dir,
        )

    def generate_formula_histogram(self, display=False, output_dir=None):
        plot_histogram(
            "formula",
            self.formula_stats,
            self.dir_path,
            display,
            output_dir,
        )

    def generate_tag_histogram(self, display=False, output_dir=None):
        plot_histogram(
            "tag",
            self.tag_stats,
            self.dir_path,
            display,
            output_dir,
        )
```

Although the code was designed to provide a simple function for each task, writing docstrings for each function has become difficult to maintain and update. This redundancy poses a challenge for code maintenance.

### Proposed solution

Pass a variable or use an `enum` to modularize the code, reducing repetition and improving maintainability.
