### 2024-10-27 4eefc65
Apply docstrings
