+++ src/cifkit/models/cif_ensemble.py
@@ -19,6 +19,35 @@ class CifEnsemble:
         preprocess=True,
         logging_enabled=False,
     ) -> None:
+        """Initialize a CifEnsemble object, containing a collection of Cif objects
+
+        Parameters
+        ----------
+        cif_dir_path : str
+            Path to the folder path containing .cif file(s).
+        add_nested_files : bool, optional
+            Option to include .cif files contained in sub-directories within cif_dir_path
+            , by default False
+        preprocess : bool, optional
+            Option to edit .cif files before initializing each .cif into Cif object,
+            by default True
+        logging_enabled : bool, optional
+            Option to log while pre-processing Cif objects, by default False
+
+        Attributes
+        ----------
+        dir_path: str
+            Path to the folder containing .cif files
+        file_paths: list[str]
+            List of file paths to .cif files
+        cifs: list[Cif]
+            List of Cif objects
+        file_count: int
+            Number of .cif files in the folder
+        logging_enabled: bool
+            Option to log while pre-processing Cif objects
+        """
+
         # Process each file, handling exceptions that may occur
         self.logging_enabled = logging_enabled
         file_paths = get_file_paths(
@@ -68,37 +97,80 @@ class CifEnsemble:
 
     @property
     def unique_formulas(self) -> set[str]:
-        """Get unique formulas from all .cif files in the folder."""
+        """Get unique formulas from all .cif files in the folder.
+
+        Returns
+        -------
+        set[str]
+            unique formulas
+
+        Examples
+        --------
+        >>> cif_ensemble.unique_formulas
+        {"EuIr2Ge2", "CeRu2Ge2", "LaRu2Ge2", "Mo"}
+        """
         return self._get_unique_property_values("formula")
 
     @property
     def unique_structures(self) -> set[str]:
-        """Get unique structures from all .cif files in the folder."""
+        """Get unique structures from all .cif files in the folder.
+
+        Examples
+        --------
+        >>> cif_ensemble.unique_structures
+        {"CeAl2Ga2", "W"}
+        """
         return self._get_unique_property_values("structure")
 
     @property
     def unique_tags(self) -> set[str]:
-        """Get unique formulas from all .cif files in the folder."""
+        """Get unique formulas from all .cif files in the folder.
+
+        Examples
+        --------
+        >>> cif_ensemble.unique_tags
+        {"hex", "rt", "rt_hex", ""}
+        """
         return self._get_unique_property_values("tag")
 
     @property
     def unique_space_group_names(self) -> set[str]:
-        """Get unique space groups from all .cif files in the folder."""
+        """Get unique space groups from all .cif files in the folder.
+        Examples
+        --------
+        >>> cif_ensemble.unique_space_group_names
+        {"I4/mmm", "Im-3m"}
+        """
         return self._get_unique_property_values("space_group_name")
 
     @property
     def unique_space_group_numbers(self) -> set[str]:
-        """Get unique space groups from all .cif files in the folder."""
+        """Get unique space groups from all .cif files in the folder.
+        Examples
+        --------
+        >>> cif_ensemble.unique_space_group_numbers
+        {139, 229}
+        """
         return self._get_unique_property_values("space_group_number")
 
     @property
     def unique_site_mixing_types(self) -> set[int]:
-        """Get unique site mixing types from all .cif files in the folder."""
+        """Get unique site mixing types from all .cif files in the folder.
+        Examples
+        --------
+        >>> cif_ensemble.unique_site_mixing_types
+        {"deficiency_without_atomic_mixing", "full_occupancy"}
+        """
         return self._get_unique_property_values("site_mixing_type")
 
     @property
[…]
