### 2024-10-27 2786a66
Apply pre-commit
