+++ news/doc-strings.rst
@@ -1,0 +1,23 @@
+**Added:**
+
+* Core docstrings to Cif and CifEnsembleo classes
+
+**Changed:**
+
+* <news item>
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* <news item>
+
+**Security:**
+
+* <news item>
