### 2024-10-27 8c61379
Add docstrings to histogram and api doc
