+++ src/cifkit/models/cif_ensemble.py
@@ -488,6 +488,21 @@ class CifEnsemble:
     # FIXME: refactor this section to maintain DRY principle
 
     def generate_structure_histogram(self, display=False, output_dir=None):
+        """Generate a histogram of the 'structure' property from CIF files.
+
+        This method creates a histogram based on the 'structure' statistics of the CIF
+        files. If 'output_dir' is specified, the histogram image (.png) will be saved to
+        that directory. If 'output_dir' is not specified, the image will be saved to the
+        directory specified by 'self.dir_path'.
+
+        Parameters
+        ----------
+        display : bool, optional
+            If True, the plot is displayed using plt.show(). Default is False.
+        output_dir : str, optional
+            The directory path where the histogram should be saved. If None,
+            the histogram is saved in the directory defined by 'self.dir_path'.
+        """
         plot_histogram(
             "structure",
             self.structure_stats,
@@ -497,6 +512,21 @@ class CifEnsemble:
         )
 
     def generate_formula_histogram(self, display=False, output_dir=None):
+        """Generate a histogram of the 'formula' property from CIF files.
+
+        This method creates a histogram based on the 'formula' statistics of the CIF
+        files. If 'output_dir' is specified, the histogram image (.png) will be saved to
+        that directory. If 'output_dir' is not specified, the image will be saved to the
+        directory specified by 'self.dir_path'.
+
+        Parameters
+        ----------
+        display : bool, optional
+            If True, the plot is displayed using plt.show(). Default is False.
+        output_dir : str, optional
+            The directory path where the histogram should be saved. If None,
+            the histogram is saved in the directory defined by 'self.dir_path'.
+        """
         plot_histogram(
             "formula",
             self.formula_stats,
@@ -506,6 +536,21 @@ class CifEnsemble:
         )
 
     def generate_tag_histogram(self, display=False, output_dir=None):
+        """Generate a histogram of the 'tag' property from CIF files.
+
+        This method creates a histogram based on the 'tag' statistics of the CIF
+        files. If 'output_dir' is specified, the histogram image (.png) will be saved to
+        that directory. If 'output_dir' is not specified, the image will be saved to the
+        directory specified by 'self.dir_path'.
+
+        Parameters
+        ----------
+        display : bool, optional
+            If True, the plot is displayed using plt.show(). Default is False.
+        output_dir : str, optional
+            The directory path where the histogram should be saved. If None,
+            the histogram is saved in the directory defined by 'self.dir_path'.
+        """
         plot_histogram(
             "tag",
             self.tag_stats,
@@ -517,6 +562,21 @@ class CifEnsemble:
     def generate_space_group_number_histogram(
         self, display=False, output_dir=None
     ):
+        """Generate a histogram of the 'space_group_number' property from CIF files.
+
+        This method creates a histogram based on the 'space_group_number' statistics of
+        the CIF files. If 'output_dir' is specified, the histogram image (.png) will be
+        saved to that directory. If 'output_dir' is not specified, the image will be saved
+        to the directory specified by 'self.dir_path'.
+
+        Parameters
+        ----------
+        display : bool, optional
+            If True, the plot is displayed using plt.show(). Default is False.
+        output_dir : str, optional
+            The directory path where the histogram should be saved. If None,
+            the histogram is saved in the directory defined by 'self.dir_path'.
+        """
         plot_histogram(
             "space_group_number",
             self.space_group_number_stats,
@@ -528,6 +588,21 @@ class CifEnsemble:
     def generate_space_group_name_histogram(
         self, display=False, output_dir=None
     ):
+        """Generate a histogram of the 'space_group_name' property from CIF files.
+
[…]
