### 2024-10-27 d18d9d4
Revert "Add docstrings to histogram and api doc"
This reverts commit 8c613798413b38ae7eb7b0b63aae9d0cf1f0e95d.
