### 2024-10-28 68fd91a
Add docstrings to src
