### 2024-10-28 eeeb4ae
Add API docs and conda install
