### 2024-10-28 e430035
Update readme with added citations
