+++ paper.md
@@ -1,5 +1,6 @@
 ---
-title: "cifkit: A user-friendly Python package for high-throughput CIF analysis"
+title:
+  "cifkit: A Python package for coordination geometry and atomic site analysis"
 tags:
   - Python
   - CIF
@@ -36,106 +37,90 @@ bibliography: paper.bib
 
 # Summary
 
-`cifkit` provides higher-level utility functions and variables from .cif files,
-which are standard file formats for storing crystallographic data such as atomic
-fractional coordinates, symmetry operations, and unit cell dimensions. `cifkit`
-serves as an engine for building Python applications that automate crystal
-structure analysis, enabling the extraction of physics-based information crucial
-for understanding geometric configurations and identifying irregularities.
-`cifkit` also offers various tools for determining coordination numbers,
-plotting coordination geometry-based polyhedron from each site, calculating bond
-fractions, moving and copying .cif files based on a set of attributes, and
-determining atomic mixing information.
+`cifkit` provides higher-level functions and properties for coordination
+geometry and atomic site analysis from .cif files, which are standard file
+formats for storing crystallographic data such as atomic fractional coordinates,
+symmetry operations, and unit cell dimensions. Designed for functionalities demanded by experimental synthesists, cifkit has been used as a backend for Python applications that automate crystal structure analysis, enabling the extraction of physics-based features crucial for understanding
+geometric configurations and identifying irregularities. `cifkit` offers
+functions such as plotting coordination geometry-based polyhedron from each
+site, calculating bond fractions, determining atomic mixing information, and
+sorting .cif files based on a set of attributes.
 
 # Statement of need
 
-In solid state chemistry and materials science, the Crystallographic Information
-File (CIF) [@hall_crystallographic_1991] is the predominant file format used to
-store and distribute crystal structure information. There are open-source Python
-packages that read, edit, and create CIF files. Python Materials Genomics
-(pymatgen) [@ong_python_2013] offers functionalities beyond the aforementioned
-features, such as generating electronic structure properties and phase diagrams.
-Similarly, the Atomic Simulation Environment (ASE) [@larsen_atomic_2017]
-provides a suite of powerful tools for generating and running atomistic
-simulations.
+In solid state chemistry and materials science, the Crystallographic Information File (CIF) [@hall_crystallographic_1991] is the primary file format for storing and distributing crystal structure information. Open-source Python packages for reading, editing, and creating CIF files include Python Materials Genomics (pymatgen) [@ong_python_2013] and the Atomic Simulation Environment (ASE) [@larsen_atomic_2017]. Pymatgen offers advanced functionalities such as generating electronic structure properties, phase diagrams, and implementing coordination environment identification through ChemEnv [@waroquiers_chemenv_2020]. ASE provides a comprehensive suite of tools for generating and running atomistic simulations.
 
-`cifkit` distinguishes itself from existing libraries by offering higher-level
-functions and variables that enable users to perform complex tasks efficiently
-with a few lines of code. `cifkit` not only facilitates the visualization of
-coordination geometry from each site but also extracts physics-based features
-like volume and packing efficiency, crucial for structural analysis in ML tasks.
-Additionally, it extracts atomic mixing information at the bond pair level—tasks
-that would otherwise require extensive manual effort using GUI-based tools like
-VESTA, Diamond, and CrystalMaker, due to the lack of readily available
-higher-level functions.
[…]
