+++ .github/ISSUE_TEMPLATE/release_checklist.md
@@ -10,13 +10,17 @@ assignees: ""
 
 - [ ] All PRs/issues attached to the release are merged.
 - [ ] All the badges on the README are passing.
-- [ ] License information is verified as correct. If you are unsure, please comment below.
-- [ ] Locally rendered documentation contains all appropriate pages, including API references (check no modules are
-  missing), tutorials, and other human written text is up-to-date with any changes in the code.
-- [ ] Installation instructions in the README, documentation and on the website (e.g., diffpy.org) are updated and
-  tested
-- [ ] Successfully run any tutorial examples or do functional testing in some other way.
+- [ ] License information is verified as correct. If you are unsure, please
+      comment below.
+- [ ] Locally rendered documentation contains all appropriate pages, including
+      API references (check no modules are missing), tutorials, and other human
+      written text is up-to-date with any changes in the code.
+- [ ] Installation instructions in the README, documentation and on the website
+      (e.g., diffpy.org) are updated and tested
+- [ ] Successfully run any tutorial examples or do functional testing in some
+      other way.
 - [ ] Grammar and writing quality have been checked (no typos).
 
-Please mention @sbillinge when you are ready for release. Include any additional comments necessary, such as
-version information and details about the pre-release.
+Please mention @sbillinge when you are ready for release. Include any additional
+comments necessary, such as version information and details about the
+pre-release.
