+++ docs/index.md
@@ -2,15 +2,14 @@
 
 ## Statement of need
 
-`cifkit` uses .cif files by offering higher-level
-functions and variables that enable users to perform complex tasks efficiently
-with a few lines of code. `cifkit` facilitates the visualization of
-coordination geometry from each site but also extracts physics-based features
-like volume and packing efficiency, crucial for structural analysis in ML tasks.
-Additionally, it extracts atomic mixing information at the bond pair level—tasks
-that would otherwise require extensive manual effort using GUI-based tools like
-VESTA, Diamond, and CrystalMaker, due to the lack of readily available
-higher-level functions.
+`cifkit` uses .cif files by offering higher-level functions and variables that
+enable users to perform complex tasks efficiently with a few lines of code.
+`cifkit` facilitates the visualization of coordination geometry from each site
+but also extracts physics-based features like volume and packing efficiency,
+crucial for structural analysis in ML tasks. Additionally, it extracts atomic
+mixing information at the bond pair level—tasks that would otherwise require
+extensive manual effort using GUI-based tools like VESTA, Diamond, and
+CrystalMaker, due to the lack of readily available higher-level functions.
 
 Further enhancing its utility, cifkit excels in sorting, preprocessing, and
 understanding the distribution of underlying CIF files. Common issues in CIF
@@ -45,7 +44,10 @@ mixing, among other parameters.
 
 ## Processing speed expectation
 
-Based on the Apple M1 iMac chip, processing a .cif file of 150 files takes around 50 seconds, but this depends on the size of the unit cell in the .cif files. Processing 10,000 .cif files or more may take about an hour, but this also depends on the laptop.
+Based on the Apple M1 iMac chip, processing a .cif file of 150 files takes
+around 50 seconds, but this depends on the size of the unit cell in the .cif
+files. Processing 10,000 .cif files or more may take about an hour, but this
+also depends on the laptop.
 
 ## Installation
 
@@ -55,7 +57,6 @@ Python 3.10, 3.11, 3.12 are supported.
 [![PyPi version](https://img.shields.io/pypi/v/cifkit.svg)](https://pypi.python.org/pypi/cifkit)
 [![Conda version](https://img.shields.io/conda/vn/conda-forge/cifkit)](https://anaconda.org/conda-forge/cifkit)
 
-
 Option 1. conda install
 
 The preferred method is to install `cifkit` using Conda.
@@ -64,7 +65,6 @@ The preferred method is to install `cifkit` using Conda.
 conda create cifkit_env cifkit
 conda activate cifkit_env
 ```
-
 
 If the above option does not work, please feel free to use pip install.
 
@@ -85,8 +85,8 @@ generates supercells, and computes nearest neighbors. It also determines
 coordination numbers using four different methods and generates polyhedrons for
 each site.
 
-The example below uses `cifkit` to visualize the polyhedron generated from each atomic site based on the
-coordination number geometry.
+The example below uses `cifkit` to visualize the polyhedron generated from each
+atomic site based on the coordination number geometry.
 
 ```python
 from cifkit import Cif
@@ -102,7 +102,9 @@ for label in site_labels:
 
 ![Polyhedron generation](assets/img/ErCoIn_polyhedron.png)
 
-For more, visit [https://bobleesj.github.io/cifkit/notebooks/01_cif](https://bobleesj.github.io/cifkit/notebooks/01_cif) for a full tutorial with an interactive Google Codelab link.
+For more, visit
[…]
