### 2024-10-28 bc4c483
remove click requirements
