+++ requirements/conda.txt
@@ -3,5 +3,3 @@ matplotlib-base
 numpy
 pyvista
 scipy
-setuptools
-click
