+++ requirements/pip.txt
@@ -3,5 +3,4 @@ matplotlib
 numpy
 pyvista
 scipy
-setuptools
-click
+
