### 2024-10-28 f65e61f
remove click support in cif ensemble
