+++ paper.md
@@ -1,3 +1,4 @@
+
 ---
 title:
   "cifkit: A Python package for coordination geometry and atomic site analysis"
@@ -171,10 +172,10 @@ testing. We acknowledge Fabian Zills for his recommendations on Python tooling.
 
 We thank the developers of the following dependencies:
 
-- gemmi [wojdyr_gemmi_2022]: .cif parsing and space group operations
+- gemmi [w@ojdyr_gemmi_2022]: .cif parsing and space group operations
 - matplotlib [@hunter_matplotlib_2007]: visualization of histograms
-- numpy [harris_array_2020]: angle conversion, linear algebra
-- pyvista [sullivan_pyvista_2019]: visualization of polyhedra
-- scipy [virtanen_scipy_2020]: minimization function to refine of CIF radius
+- numpy [@harris_array_2020]: angle conversion, linear algebra
+- pyvista [@sullivan_pyvista_2019]: visualization of polyhedra
+- scipy [@virtanen_scipy_2020]: minimization function to refine of CIF radius
 
 # References
