+++ requirements/pip.txt
@@ -3,4 +3,3 @@ matplotlib
 numpy
 pyvista
 scipy
-
