+++ paper-review-2.md
@@ -1,0 +1,161 @@
+## Review by @lancekavalsky
+
+**Code comments:**
+
+> For io related methods, e.g. the ones that generate histograms, more clarity is needed regarding where they write things by default. Running the histogram tutorial from the documentation, it wrote the histograms to a folder deep in my conda environment site-packages, which would likely not be intuitive to many users (particularly as this package is presented as being catered to users with less coding experience) and may cause issues on shared resources.
+
+Thank you for pointing that out. Yes, the histogram images were saved to the Anaconda environment, as that was where the .cif files were provided by default in the package. In the docs, I have included information about options to set the output path for users:
+
+```python
+# Optional: Specify the output directory where the .png file will be saved.
+ensemble.generate_site_mixing_type_histogram(output_dir="path/to/directory")
+
+# Optional: Call plt.show() to display the histogram on screen.
+ensemble.generate_site_mixing_type_histogram(display=False)
+```
+
+For API users, I have include numpy-stlye docstrings:
+
+```python
+def generate_supercell_size_histogram(
+    self, display=False, output_dir=None
+):
+    """Generate a histogram of the 'supercell_count' property from CIF files.
+
+    This method creates a histogram based on the 'supercell_count' statistics of
+    the CIF files. If 'output_dir' is specified, the histogram image (.png) will be
+    saved to that directory. If 'output_dir' is not specified, the image will be saved
+    to the directory specified by 'self.dir_path'.
+
+    Parameters
+    ----------
+    display : bool, optional
+        If True, the plot is displayed using plt.show(). Default is False.
+    output_dir : str, optional
+        The directory path where the histogram should be saved. If None,
+        the histogram is saved in the directory defined by 'self.dir_path'.
+    """
+```
+
+> I will echo the previous comment that more docstrings would be invaluable in helping code clarity. In particular, I would urge adopting a standardized approach to this, such as Numpy, to be more in line with community standards
+
+Per your suggestion, I have used Numpy docstrings to the core `Cif` and `CifEnsemble` classes. I will continue to udpate the docs after I make someo f the functions more dmoularlized like combinignign generating histogram functions into a singlo to avoid verbosity. https://github.com/bobleesj/cifkit/pull/54
+
+> I have opened a technical issue I encountered here
+
+Thank you. I have addressed the issue via this PR: https://github.com/bobleesj/cifkit/pull/47. To ensure compatibility, I created a test function to ensure that raw .cif files sourced from ICSD, COD can be all parsed and the supercell can be generated:
+
+```python
+@pytest.mark.parametrize(
+    "cif_folder_path, expected_file_count, expected_supercell_stats",
+    [
+        ("tests/data/cif/sources/ICSD", 4, {216: 2, 307: 1, 320: 1}),
+        ("tests/data/cif/sources/COD", 2, {519: 1, 1383: 1}),
+        ("tests/data/cif/sources/MP", 2, {108: 1, 594: 1}),
+        ("tests/data/cif/sources/PCD", 1, {364: 1}),
+        ("tests/data/cif/sources/MS", 1, {2988: 1}),
+        ("tests/data/cif/sources/CCDC", 1, {3844: 1}),
+    ],
+)
+```
+
+**Documentation comments:**
+
+> As the core classes for this package are Cif and CifEnsemble, more explicit explanations as to the inputs and parameters would be helpful -- especially for CifEnsemble. For example, unless I'm missing it, a comprehensive list of everything that the preprocess parameter triggers is not mentioned in the documentation.
+
[…]
