+++ paper-review-1.md
@@ -1,0 +1,82 @@
+## Review by @espottesmith
+
+**Paper comments:**
+
+> The summary is generally appropriate. From context, a non-expert might realize what *.cif files are (since crystal structures are mentioned in the second sentence), but this could be more explicitly stated in the first sentence.
+
+Thank you for your suggestion. The summary begins with
+
+```
+`cifkit` provides higher-level functions and properties for coordination
+geometry and atomic site analysis from .cif files, which are standard file
+formats for storing crystallographic data such as atomic fractional coordinates,
+symmetry operations, and unit cell dimensions.
+```
+
+> I come from a computational background, so I'm not the target audience, but I am unconvinced that the basic functionality (i.e., parsing CIF files and extracting information like structures and formulas) is easier in cifkit than in pymatgen/ASE. Similar numbers of lines of code would be required, at least comparing pymatgen and the example given in the cifkit paper (see https://pymatgen.org/usage.html#reading-and-writing-structuresmolecules). Considering higher-level features of cifkit, including filtering groups of CIFs and visualization, I think that cifkit distinguishes itself from other codes. In this area, the authors' argument is more convincing. Perhaps a slight reframing of the Statement of Need is appropriate?
+
+This is a valid point. I have modified the "Statement of Need" and emphaszie the utility of higher-level functions by modifying as follow:
+
+```
+`cifkit` distinguishes itself from existing libraries by offering higher-level
+functions and variables that allow solid-state synthesists to obtain intuitive,
+yet physically impactful properties. It facilitates the visualization of
+coordination geometry from each site using four coordination determination
+methods and extracts physics-based features like volume and packing
+efficiency—crucial for structural analysis in machine learning tasks. Moreover,
+`cifkit` extracts atomic mixing information at the bond pair level, tasks that
+would otherwise require extensive manual effort using GUI-based tools like
+VESTA, Diamond, and CrystalMaker. These functions can be further developed
+on-demand, as demonstrated by `cifkit`'s ability to extract coordination
+geometry information based on four coordination number determination methods for
+a newly discovered phase [@tyvanchuk_crystal_2024].
+```
+
+> Otherwise, the paper looks good!
+
+Thank you!
+
+**Code comments:**
+
+> The folder "[occupacny"](https://github.com/bobleesj/cifkit/tree/main/src/cifkit/occupacny) should probably be "occupancy"
+
+Modified. Thank you.
+
+> I initially tried installing with Python 3.9 (as far as I can tell, the "pyproject.toml" and the docs don't specify a version). This caused problems, as cifkit is using type annotation syntax that is not supported in Python 3.9. I now see on PYPI that cifkit only supports Python 3.10, 3.11, and 3.12. Please make this more clear for users.
+
+In `pyproject.toml`, I have added the following `requires-python = '>=3.10, <3.13` to support 3.10, 3.11, 3.12. Per your suggestion, I have included the supported Python badged in the documentation under the Getting Started page.
+
+> In terms of code style, more extensive docstrings (for instance, explaining what the inputs and outputs should be) would be appreciated. It seems there are already some "TODO"s to this effect (see e.g., models/cif.py), so perhaps the authors are planning on making these changes soon. Function bodies are typically well commented (note that I have not read every file)
+
+I have added numpy-style docstrings for the main classes `Cif` and `CifEnsemble` here: https://github.com/bobleesj/cifkit/pull/49. I've also identified some code methods that could be refactored, as noted here: https://github.com/bobleesj/cifkit/issues/53. My next step is to further refine these docstrings by consolidating them into reusable methods.
+
+**Documentation comments:**
+
[…]
