+++ paper-review-2.md
@@ -1,10 +1,10 @@
-## Review by @lancekavalsky
+## Response to @lancekavalsky
 
 **Code comments:**
 
 > For io related methods, e.g. the ones that generate histograms, more clarity is needed regarding where they write things by default. Running the histogram tutorial from the documentation, it wrote the histograms to a folder deep in my conda environment site-packages, which would likely not be intuitive to many users (particularly as this package is presented as being catered to users with less coding experience) and may cause issues on shared resources.
 
-Thank you for pointing that out. Yes, the histogram images were saved to the Anaconda environment, as that was where the .cif files were provided by default in the package. In the docs, I have included information about options to set the output path for users:
+Thank you for the details. Yes, the histogram images were saved to the Anaconda environment, as that was where the .cif files were provided by default in the installed package. In the docs, I have included information about options to set the output path for users for clarity:
 
 ```python
 # Optional: Specify the output directory where the .png file will be saved.
@@ -14,7 +14,7 @@ ensemble.generate_site_mixing_type_histogram(output_dir="path/to/directory")
 ensemble.generate_site_mixing_type_histogram(display=False)
 ```
 
-For API users, I have include numpy-stlye docstrings:
+For API doc users, I have included docstrings:
 
 ```python
 def generate_supercell_size_histogram(
@@ -39,11 +39,11 @@ def generate_supercell_size_histogram(
 
 > I will echo the previous comment that more docstrings would be invaluable in helping code clarity. In particular, I would urge adopting a standardized approach to this, such as Numpy, to be more in line with community standards
 
-Per your suggestion, I have used Numpy docstrings to the core `Cif` and `CifEnsemble` classes. I will continue to udpate the docs after I make someo f the functions more dmoularlized like combinignign generating histogram functions into a singlo to avoid verbosity. https://github.com/bobleesj/cifkit/pull/54
+Per your suggestion, I have added NumPy-style docstrings to the core `Cif` and `CifEnsemble` classes. I will continue updating the documentation after modularizing some functions, such as combining the histogram generation functions into a single one to reduce verbosity.
 
 > I have opened a technical issue I encountered here
 
-Thank you. I have addressed the issue via this PR: https://github.com/bobleesj/cifkit/pull/47. To ensure compatibility, I created a test function to ensure that raw .cif files sourced from ICSD, COD can be all parsed and the supercell can be generated:
+Thank you. I have addressed the issue via this PR: https://github.com/bobleesj/cifkit/pull/47. To ensure compatibility, I created a test function to ensure that raw .cif files sourced from ICSD, COD can be parsed and the supercell can be generated:
 
 ```python
 @pytest.mark.parametrize(
@@ -63,7 +63,7 @@ Thank you. I have addressed the issue via this PR: https://github.com/bobleesj/c
 
 > As the core classes for this package are Cif and CifEnsemble, more explicit explanations as to the inputs and parameters would be helpful -- especially for CifEnsemble. For example, unless I'm missing it, a comprehensive list of everything that the preprocess parameter triggers is not mentioned in the documentation.
 
-Thank you. I have included docstrings for Cif and CifEnsemble classes to provide explicit parameters as shown below and the preprocessing triggers in the docstrings. The docs can be further improved. I hope it serves the purpose for now. I plan on maintaining this package during my graduate school.
+Thank you. I have included docstrings for the `Cif` and `CifEnsemble` classes, providing explicit parameters and preprocessing triggers as shown below. While the documentation can be further refined, I believe it serves its purpose for now.
 
 ```python
 class CifEnsemble:
@@ -115,7 +115,7 @@ class CifEnsemble:
 
[…]
