### 2024-10-28 24c26db
Fix typo in paper.md
