### 2024-10-28 41b8ed4
Run news on pull request only
