+++ .github/workflows/check-news-item.yml
@@ -2,8 +2,6 @@ name: Check for News
 
 on:
   pull_request_target:
-    branches:
-    - main
 
 jobs:
   build:
