### 2024-10-28 a6fbcd7
Fix typo with pre-commit codespell
