+++ docs/index.md
@@ -4,7 +4,7 @@
 
 `cifkit` uses .cif files by offering higher-level functions and variables that
 enable users to perform complex tasks efficiently with a few lines of code. `cifkit` distinguishes itself from existing libraries by offering higher-level
-functions and variables that allow solid-state synthesists to obtain intuitive and 
+functions and variables that allow solid-state synthesists to obtain intuitive and
 measurable properties impactful properties. It facilitates the visualization of
 coordination geometry from each site using four coordination determination
 methods and extracts physics-based features like volume and packing
