+++ .codespell/ignore_lines.txt
@@ -1,22 +1,0 @@
-;; Please include filenames and explanations for each ignored line.
-;; See https://docs.openverse.org/meta/codespell.html for docs.
-
-;; .github/workflows/draft_releases.yml
-;; The release-drafter/release-drafter configuration variable name cannot be changed
-          commitish: main
-
-;; frontend/src/locales/scripts/en.json5
-;; Prettier insists we escape a single quote rather than the double quotes and codespell
-;; does not understand the escaped `\'t` as "couldn't". It instead just sees "couldn".
-    heading: 'We couldn\'t find anything for "{query}".',
-
-;; catalog/tests/dags/providers/provider_api_scripts/test_wikimedia_commons.py
-;; "Titel" matches "title", but the phrase is in Dutch, not English, so "titel"
-;; is actually correct. Similarly, "als" incorrectly matches to "also" in the same
-;; block of Dutch text.
-        "Identificatie Titel(s):  Allegorie op kunstenaar Francesco Mazzoli, "
-        "bekend als Parmigianino"
-
-;; packages/js/eslint-plugin/configs/vue.ts
-;; `te` gets matched with `the` and others
-const i18nDestructureRules = ["t", "tc", "te", "td", "d", "n"].map(
