### 2024-10-28 ad316e3
fix install by adding -n to conda
