### 2024-10-28 7588c54
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/pre-commit/pre-commit-hooks: v4.6.0 → v5.0.0](https://github.com/pre-commit/pre-commit-hooks/compare/v4.6.0...v5.0.0)
- [github.com/pycqa/flake8: 7.0.0 → 7.1.1](https://github.com/pycqa/flake8/compare/7.0.0...7.1.1)
- [github.com/pre-commit/pre-commit-hooks: v4.4.0 → v5.0.0](https://github.com/pre-commit/pre-commit-hooks/compare/v4.4.0...v5.0.0)
- [github.com/codespell-project/codespell: v2.2.4 → v2.3.0](https://github.com/codespell-project/codespell/compare/v2.2.4...v2.3.0)
