+++ news/doc-strings.rst
@@ -1,23 +1,0 @@
-**Added:**
-
-* Core docstrings to Cif and CifEnsembleo classes
-
-**Changed:**
-
-* <news item>
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* <news item>
-
-**Fixed:**
-
-* <news item>
-
-**Security:**
-
-* <news item>
