+++ news/pre-commit.rst
@@ -1,24 +1,0 @@
-**Added:**
-
-* pre-commit hook for each pull request
-* automated PyPI
-
-**Changed:**
-
-* <news item>
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* <news item>
-
-**Fixed:**
-
-* Simpler issue templates used in Billinge Group
-
-**Security:**
-
-* <news item>
