+++ news/lint.rst
@@ -1,23 +1,0 @@
-**Added:**
-
-* <news item>
-
-**Changed:**
-
-* <news item>
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* Support for click module
-
-**Fixed:**
-
-* <news item>
-
-**Security:**
-
-* <news item>
