+++ news/workflow.rst
@@ -1,23 +1,0 @@
-**Added:**
-
-* Use GitHub Actions to deploy to PyPI and update CHANGELOG.rst
-
-**Changed:**
-
-* <news item>
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* <news item>
-
-**Fixed:**
-
-* <news item>
-
-**Security:**
-
-* <news item>
