### 2024-10-28 0f8ce4b
Fix image link to polyhedron
