### 2024-10-28 2a2473d
Fix typo in .png file path
