### 2024-10-28 bobleesj/cifkit#27 @bobleesj
Thank you. All of the points have been addressed.
1.

```python
from cifkit import CifEnsemble, Cif

ensemble_CCDC = CifEnsemble("tests/data/cif/sources/CCDC")
ensemble_COD = CifEnsemble("tests/data/cif/sources/COD")
ensemble_ICSD = CifEnsemble("tests/data/cif/sources/ICSD")
ensemble_MP = CifEnsemble("tests/data/cif/sources/MP")
ensemble_PCD = CifEnsemble("tests/data/cif/sources/PCD")
ensemble_MS = CifEnsemble("tests/data/cif/sources/MS")

print("ensemble_CCDC:", ensemble_CCDC.CN_unique_values_by_best_methods)
print("ensemble_COD:", ensemble_COD.CN_unique_values_by_best_methods)
print("ensemble_ICSD:", ensemble_ICSD.CN_unique_values_by_best_methods)
print("ensemble_MP:", ensemble_MP.CN_unique_values_by_best_methods)
print("ensemble_PCD:", ensemble_PCD.CN_unique_values_by_best_methods)
print("ensemble_MS:", ensemble_MS.CN_unique_values_by_best_methods)

ensemble_CCDC: {16, 11, 12, 13, 14, 15}
ensemble_COD: {10, 4, 7, 8, 682}
ensemble_ICSD: {6, 7, 12, 13}
ensemble_MP: {4, 6}
ensemble_PCD: {6}
ensemble_MS: {19, 9, 10, 11, 12, 13, 14}
```

2. docstrings are provided through the following PRs https://github.com/bobleesj/cifkit/pull/54, https://github.com/bobleesj/cifkit/pull/49https://github.com/bobleesj/cifkit/pull/49, and https://github.com/bobleesj/cifkit/pull/55 and they are available under https://bobleesj.github.io/cifkit/api/

3. removed pinning. Thanks!
