### 2024-10-28 07b0ebb
Build jupter doc outputs
