### 2024-10-28 575d246
Remove update section
