+++ paper-review-2.md
@@ -2,9 +2,17 @@
 
 **Code comments:**
 
-> For io related methods, e.g. the ones that generate histograms, more clarity is needed regarding where they write things by default. Running the histogram tutorial from the documentation, it wrote the histograms to a folder deep in my conda environment site-packages, which would likely not be intuitive to many users (particularly as this package is presented as being catered to users with less coding experience) and may cause issues on shared resources.
+> For io related methods, e.g. the ones that generate histograms, more clarity
+> is needed regarding where they write things by default. Running the histogram
+> tutorial from the documentation, it wrote the histograms to a folder deep in
+> my conda environment site-packages, which would likely not be intuitive to
+> many users (particularly as this package is presented as being catered to
+> users with less coding experience) and may cause issues on shared resources.
 
-Thank you for the details. Yes, the histogram images were saved to the Anaconda environment, as that was where the .cif files were provided by default in the installed package. In the docs, I have included information about options to set the output path for users for clarity:
+Thank you for the details. Yes, the histogram images were saved to the Anaconda
+environment, as that was where the .cif files were provided by default in the
+installed package. In the docs, I have included information about options to set
+the output path for users for clarity:
 
 ```python
 # Optional: Specify the output directory where the .png file will be saved.
@@ -37,13 +45,21 @@ def generate_supercell_size_histogram(
     """
 ```
 
-> I will echo the previous comment that more docstrings would be invaluable in helping code clarity. In particular, I would urge adopting a standardized approach to this, such as Numpy, to be more in line with community standards
+> I will echo the previous comment that more docstrings would be invaluable in
+> helping code clarity. In particular, I would urge adopting a standardized
+> approach to this, such as Numpy, to be more in line with community standards
 
-Per your suggestion, I have added NumPy-style docstrings to the core `Cif` and `CifEnsemble` classes. I will continue updating the documentation after modularizing some functions, such as combining the histogram generation functions into a single one to reduce verbosity.
+Per your suggestion, I have added NumPy-style docstrings to the core `Cif` and
+`CifEnsemble` classes. I will continue updating the documentation after
+modularizing some functions, such as combining the histogram generation
+functions into a single one to reduce verbosity.
 
 > I have opened a technical issue I encountered here
 
-Thank you. I have addressed the issue via this PR: https://github.com/bobleesj/cifkit/pull/47. To ensure compatibility, I created a test function to ensure that raw .cif files sourced from ICSD, COD can be parsed and the supercell can be generated:
+Thank you. I have addressed the issue via this PR:
+https://github.com/bobleesj/cifkit/pull/47. To ensure compatibility, I created a
+test function to ensure that raw .cif files sourced from ICSD, COD can be parsed
+and the supercell can be generated:
 
 ```python
 @pytest.mark.parametrize(
@@ -61,9 +77,16 @@ Thank you. I have addressed the issue via this PR: https://github.com/bobleesj/c
 
 **Documentation comments:**
 
-> As the core classes for this package are Cif and CifEnsemble, more explicit explanations as to the inputs and parameters would be helpful -- especially for CifEnsemble. For example, unless I'm missing it, a comprehensive list of everything that the preprocess parameter triggers is not mentioned in the documentation.
+> As the core classes for this package are Cif and CifEnsemble, more explicit
+> explanations as to the inputs and parameters would be helpful -- especially
+> for CifEnsemble. For example, unless I'm missing it, a comprehensive list of
+> everything that the preprocess parameter triggers is not mentioned in the
+> documentation.
 
[…]
