+++ paper-review-1.md
@@ -2,9 +2,12 @@
 
 **Paper comments:**
 
-> The summary is generally appropriate. From context, a non-expert might realize what *.cif files are (since crystal structures are mentioned in the second sentence), but this could be more explicitly stated in the first sentence.
+> The summary is generally appropriate. From context, a non-expert might realize
+> what \*.cif files are (since crystal structures are mentioned in the second
+> sentence), but this could be more explicitly stated in the first sentence.
 
-Thank you for your suggestion. The summary now incorporates what *.cif files are:
+Thank you for your suggestion. The summary now incorporates what \*.cif files
+are:
 
 ```markdown
 `cifkit` provides higher-level functions and properties for coordination
@@ -13,15 +16,25 @@ formats for storing crystallographic data such as atomic fractional coordinates,
 symmetry operations, and unit cell dimensions.
 ```
 
-> I come from a computational background, so I'm not the target audience, but I am unconvinced that the basic functionality (i.e., parsing CIF files and extracting information like structures and formulas) is easier in cifkit than in pymatgen/ASE. Similar numbers of lines of code would be required, at least comparing pymatgen and the example given in the cifkit paper (see https://pymatgen.org/usage.html#reading-and-writing-structuresmolecules). Considering higher-level features of cifkit, including filtering groups of CIFs and visualization, I think that cifkit distinguishes itself from other codes. In this area, the authors' argument is more convincing. Perhaps a slight reframing of the Statement of Need is appropriate?
+> I come from a computational background, so I'm not the target audience, but I
+> am unconvinced that the basic functionality (i.e., parsing CIF files and
+> extracting information like structures and formulas) is easier in cifkit than
+> in pymatgen/ASE. Similar numbers of lines of code would be required, at least
+> comparing pymatgen and the example given in the cifkit paper (see
+> https://pymatgen.org/usage.html#reading-and-writing-structuresmolecules).
+> Considering higher-level features of cifkit, including filtering groups of
+> CIFs and visualization, I think that cifkit distinguishes itself from other
+> codes. In this area, the authors' argument is more convincing. Perhaps a
+> slight reframing of the Statement of Need is appropriate?
 
-This is a valid point. I have modified the "Statement of Need" and emphasized the utility of higher-level functions by modifying the section as follow:
+This is a valid point. I have modified the "Statement of Need" and emphasized
+the utility of higher-level functions by modifying the section as follow:
 
 ```markdown
 `cifkit` distinguishes itself from existing libraries by offering higher-level
-functions and variables that allow solid-state synthesists to obtain intuitive and
-measurable properties impactful properties. It facilitates the visualization of
-coordination geometry from each site using four coordination determination
+functions and variables that allow solid-state synthesists to obtain intuitive
+and measurable properties impactful properties. It facilitates the visualization
+of coordination geometry from each site using four coordination determination
 methods and extracts physics-based features like volume and packing
 efficiency—crucial for structural analysis in machine learning tasks. Moreover,
 `cifkit` extracts atomic mixing information at the bond pair level, tasks that
@@ -38,45 +51,88 @@ Thank you!
 
 **Code comments:**
 
-> The folder "[occupacny"](https://github.com/bobleesj/cifkit/tree/main/src/cifkit/occupacny) should probably be "occupancy"
+> The folder
+> "[occupacny"](https://github.com/bobleesj/cifkit/tree/main/src/cifkit/occupacny)
+> should probably be "occupancy"
 
 Modified. Thank you.
 
[…]
