### 2024-10-28 c733669
Update paper.md bib typo
