### 2024-10-28 #7205 @bobleesj
@rkurchin How can I re-render the manuscript here? It's in the `main` branch of https://github.com/bobleesj/cifkit/. Thanks!
