### 2024-11-05 0a3f1ea
Fix title position in joss paper.jd
