### 2024-11-05 54f187a
Use the latest udpate affiliation
