### 2024-11-10 #7205 @ml-evs
I am happy that my comments have been addressed, happy to recommend this for acceptance without needing another look. The API documentation is now clear and the statement of need better reflects the functionality of the package. Thanks for the hard work!

Just a few of final minor points:

- the grammar in the sentence L56-57 needs to be reviewed: `...and national laboratories for crystal
structural analysis and machine learning studies, and expanding.`.
- the code sample in the paper should be adjusted to fit within the margins (L54)
- I tried some more random CIF files from CSD/ICSD/MP etc. and all now work without any errors (:confetti_ball:), but some functionality that I expected to work did not. For example, in cases where the CIF only provides `chemical_formula_sum` and not `chemical_formula_structural`, the histogram generation for an formula/structure etc. for the ensemble fails. Similarly, when trying to generate polyhedra for these CIFs, the program hangs in cases where invalid input is provided (i.e., not providing a label). You could say these are either user errors or data source errors, but I think for a package aimed at non-specialists it would be helpful to provide a clear error message and fail early in cases where this info is missing. I don't see this as blocking for the paper.
