### 2024-11-10 d524f78
Address reviewer feeback on margin for code example
