### 2024-11-10 #7205 @bobleesj
Response to Nov 10, 2024 @ml-evs

> the grammar in the sentence L56-57 needs to be reviewed: ...and national laboratories for crystal structural analysis and machine learning studies, and expanding..

Right. I removed ", and expanding": `cifkit has been used for research conducted at academic and national laboratories for crystal structure analysis and machine learning studies.`

> the code sample in the paper should be adjusted to fit within the margins (L54)

Right, I re-rendered and double-checked using `>>> ensemble.filter_by_site_mixing_types(["deficiency_without_atomic_mixing"])`.

> I tried some more random CIF files from CSD/ICSD/MP etc. and all now work without any errors (:confetti_ball:), but some functionality that I expected to work did not.

Thank you for going above and beyond. The reviewers' feedback has been extremely helpful in raising the standard of this package and contribution to the broader community. I've created an issue in the repo above, and I will address this!

@lancekavalsky Would highly appreciate if you could provide some feedback :)
