### 2024-11-10 9493dce
Make filtering comment consistent in paper.md
