### 2024-11-10 #7205 @lancekavalsky
All of my comments have been comprehensively addressed and I am happy to recommend acceptance! The reworked documentation is clear (docs webpage, docstrings, overall organization, etc..) and I am no longer encountering the technical issue mentioned before.

A few final (very) minor comments regarding the paper. First, Figure 1 of the paper does not appear to be mentioned in the text. Also, the caption should be clarified to highlight that the left figure corresponds to a single cif file (as in the first example) whereas the right figure considers an ensemble of cif files (as in the second example). Lastly, for consistency in formatting the examples, the `>>>` needs to be added in front of the import lines in both examples. These are all small changes and I don't need another look.

Thanks @bobleesj for your hard work and detailed response!
