### 2024-11-11 8ad02f0
Clarify Figure 1 description and refer to it under `Examples` section
