### 2024-11-11 #7205 @bobleesj
Thank you! @lancekavalsky

> First, Figure 1 of the paper does not appear to be mentioned in the text.

Right, I've added the following two sentecnes after, `...complex tasks—such as extracting geometry-based
polyhedra descriptors from atomic sites.`

```
As shown in Figure 1, `cifkit` provides
a higher-level function to visualize the atomic site coordination geometry from a single .cif file
using the `Cif` object. It also provides an overview of multiple .cif files through the the `CifEnsemble` object.
```

> Also, the caption should be clarified to highlight that the left figure corresponds to a single cif file (as in the first example) whereas the right figure considers an ensemble of cif files (as in the second example).

Right. It wasn't clear.  The new version is `Atomic site coordination geometry from a single .cif file (left) and distribution of coordination numbers obtained from an ensemble of .cif files (right)`.

> Lastly, for consistency in formatting the examples, the >>> needs to be added in front of the import lines in both examples. These are all small changes and I don't need another look.

Right,  `>>> from cifkit import Cif, Example` and `>>> from cifkit import CifEnsemble, Example` added.

Thank you so much!

@rkurchin I think I've addressed all of the comments from the three reviewers. Thank you for coordinating!
