### 2024-11-11 #7205 @rkurchin
Editorial stuff:
* line 15: either need an "a" before polyhedron or to change it to "polyhedra"
* 19: hyphenate "solid-state"
* 42: it seems like "error handling" might be a better phrase here?

References:
* in the reference to the Tyvanchuk paper, it looks like you could use some additional `{}` around elemental symbols in the parentheses
