### 2024-11-11 5cbf60a
Add {} around {Gd}–{Tm}, {Lu}
