+++ paper.bib
@@ -5,8 +5,8 @@
 	issn = {0925-8388},
 	url = {https://www.sciencedirect.com/science/article/pii/S0925838823045449},
 	doi = {10.1016/j.jallcom.2023.173241},
-	shorttitle = {The crystal and electronic structure of \textit{{RE}}23Co6.7In20.3 (\textit{{RE}} = Gd–Tm, Lu)},
-	abstract = {New ternary rare-earth indides {RE}23Co6.7In20.3 ({RE} = Gd–Tm, Lu) have been synthesized by arc-melting the elements under argon and subsequent annealing at 870 K for 1200 h. Single-crystal X-ray diffraction revealed Er23Co6.7In20.3 to crystallize in a new structure type in {oP}100, space group Pbam and Wyckoff sequence h11g13da with a = 23.203(5), b = 28.399(5), c = 3.5306(6) Å. The crystal structures of {RE}23Co6.7In20.3 ({RE} = Tb, Ho, Er and Tm) were determined from single crystal and powder X-ray diffraction data and further investigated by {DFT} methods. The compounds belong to a large family of ternary rare-earth indides with intergrowth of the {AlB}2- and {CsCl}-type related slabs. In the Er23Co6.7In20.3 structure, four types of fragments {REIn} and {RET} of {CsCl}-type, as well as {RET}2 and {REIn}2 of {AlB}2-type, are present simultaneously. A simple Python tool was developed to determine the coordination number for each crystallographic site with various methods and tested on the complex structure of {RE}23Co6.7In20.3.},
+	shorttitle = {The crystal and electronic structure of \textit{{RE}}23Co6.7In20.3 (\textit{{RE}} = {Gd}–{Tm}, {Lu})},
+	abstract = {New ternary rare-earth indides {RE}23Co6.7In20.3 ({RE} = {Gd}–{Tm}, {Lu}) have been synthesized by arc-melting the elements under argon and subsequent annealing at 870 K for 1200 h. Single-crystal X-ray diffraction revealed Er23Co6.7In20.3 to crystallize in a new structure type in {oP}100, space group Pbam and Wyckoff sequence h11g13da with a = 23.203(5), b = 28.399(5), c = 3.5306(6) Å. The crystal structures of {RE}23Co6.7In20.3 ({RE} = Tb, Ho, Er and Tm) were determined from single crystal and powder X-ray diffraction data and further investigated by {DFT} methods. The compounds belong to a large family of ternary rare-earth indides with intergrowth of the {AlB}2- and {CsCl}-type related slabs. In the Er23Co6.7In20.3 structure, four types of fragments {REIn} and {RET} of {CsCl}-type, as well as {RET}2 and {REIn}2 of {AlB}2-type, are present simultaneously. A simple Python tool was developed to determine the coordination number for each crystallographic site with various methods and tested on the complex structure of {RE}23Co6.7In20.3.},
 	pages = {173241},
 	journaltitle = {Journal of Alloys and Compounds},
 	shortjournal = {Journal of Alloys and Compounds},
