### 2024-11-11 8cd9377
Add {Gd}–{Tm}, {Lu} to the title as well
