+++ paper.bib
@@ -1,6 +1,6 @@
 
 @article{tyvanchuk_crystal_2024,
-	title = {The crystal and electronic structure of \textit{{RE}}23Co6.7In20.3 (\textit{{RE}} = Gd–Tm, Lu): A new structure type based on intergrowth of {AlB}2- and {CsCl}-type related slabs},
+	title = {The crystal and electronic structure of \textit{{RE}}23Co6.7In20.3 (\textit{{RE}} = {Gd}–{Tm}, {Lu}): A new structure type based on intergrowth of {AlB}2- and {CsCl}-type related slabs},
 	volume = {976},
 	issn = {0925-8388},
 	url = {https://www.sciencedirect.com/science/article/pii/S0925838823045449},
