### 2024-11-11 #7205 @bobleesj
@rkurchin

- Line 15: I added "a" before "polyhedron" as suggested. Thank you.
- Line 19: I hyphenated "solid-state."
- Line 42: I replaced it with "preprocessing."
- Reference to Tyvanchuk paper: I added additional {} around elemental symbols. Rendered correctly. Thanks for catching that.
