### 2024-11-11 21433eb
Add news and remove node js files
