+++ news/JOSS.rst
@@ -1,0 +1,23 @@
+**Added:**
+
+* cifkit accepted to JOSS
+
+**Changed:**
+
+* <news item>
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* <news item>
+
+**Security:**
+
+* <news item>
