+++ news/JOSS.rst
@@ -1,23 +1,0 @@
-**Added:**
-
-* cifkit accepted to JOSS
-
-**Changed:**
-
-* <news item>
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* <news item>
-
-**Fixed:**
-
-* <news item>
-
-**Security:**
-
-* <news item>
