### 2024-11-11 #7205 @bobleesj
### Additional Author Tasks After Review is Complete
- [x] Make a release of the software with the latest changes from the review and post the version number here.
- [x] This is the version that will be used in the JOSS paper.

v1.0.6 is the version! https://github.com/bobleesj/cifkit/releases/tag/1.0.6

https://doi.org/10.5281/zenodo.14080155

- [x] Make sure that the license listed for the archive is the same as the software license

@rkurchin  Done!
