### 2024-11-12 #7205 @rkurchin
Please make sure:
* the author list in the archive matches the manuscript
* the license in the archive matches the one in the repository
