### 2024-11-12 #7205 @bobleesj
License is double-checked. @rkurchin The software is primarily built and maintained by me. Hence, I am currently the single author in `pyprojet.toml`. Would you also need the second author of the manuscript (Anton Oliynyk)  included in the zenodo archive?
