### 2024-11-12 #7205 @rkurchin
@bobleesj can you clarify what the contributions of Anton Oliynyk were in that case?

We usually do want the authors of the archive and the paper to match if possible, though I believe there can be exceptions made; @openjournals/bcm-eics can perhaps jump in in that regard.
