### 2024-11-12 #7205 @bobleesj
Anton Oliynyk's contributions include not limited to conceptually designing the specific methods embedded in the applications mentioned in the manuscript. He also validated these functions using GUI-based software and also provided feedback when new bugs are encountered and how to navigate the problem together.
