### 2024-11-13 #7205 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
