### 2024-11-15 #7205 @Kevin-Mattheus-Moerman
@bobleesj as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks I'm processing, some of which may need your attention.

- [ ] Please ensure the archive listed license tag matches your software release tag: `1.0.6`. You do not need to create a new archive link, you should be able to manually edit the tag when you log into ZENODO.
- [ ] Please spell out country names in your affiliations, so do not use the acronym `USA`.
