### 2024-11-15 #7205 @bobleesj
@Kevin-Mattheus-Moerman Thank you!

>  Please ensure the archive listed license tag matches your software release tag: 1.0.6. You do not need to create a new archive link, you should be able to manually edit the tag when you log into ZENODO.

Now, it is updated:

```
Version 1.0.6
[10.5281/zenodo.14086943](https://doi.org/10.5281/zenodo.14086943)
```

>  Please spell out country names in your affiliations, so do not use the acronym USA.

Now `paper.md` writes:

```
affiliations:
  - name:
      Department of Applied Physics and Applied Mathematics, Columbia
      University, New York, NY 10027, United States
    index: 1
  - name:
      Department of Chemistry, Hunter College, City University of New York, New
      York, NY 10065, United States
    index: 2
  - name:
      Ph.D. Program in Chemistry, The Graduate Center of the City University of
      New York, New York, NY 10016, United States
```
